+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Change Properties Provider to be a Converter (#4666)
 - Define a type `WatcherFunc` for onChange func instead of func pointer (#4656)
 - Remove deprecated `configtest.LoadConfig` and `configtest.LoadConfigAndValidate` (#4659)
+- Replace service.ConfigMapConverterFunc with config.MapConverter interface (#4668)
 
 ## 💡 Enhancements 💡